requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
ecdsa==0.17.0
fastapi==0.67.0
h11==0.12.0
jsonschema==3.2.0
ocpp==0.8.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyrsistent==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
websockets==9.1
twilio==7.4.0
boto3
schedule==1.1.0
arq
transitions
python-dotenv
loguru
fastapi_sso
aiohttp
simplejson